arrange files of creating reports #47
