+++ src/controllers/reports.js
@@ -2,6 +2,10 @@ const convertToCamelCase = require('camelcase-keys-deep');
 const convertToSnakeCase = require('snakecase-keys');
 const bonusQuery = require('../database/query/bonusQuery');
 const deductionsQuery = require('../database/query/deductionsQuery');
+const exemptions = require('../database/query/exemptions');
+const deductionOperations = require('./helpers/deductionsOperations');
+const exemptionsOperations = require('./helpers/exemptionsOperations');
+
 const {
   certificate,
   employee,
@@ -10,7 +14,6 @@ const {
   purchasesEmployee,
   report,
 } = require('../database/models');
-const { bonus, deductions } = require('../functions/index');
 
 exports.get = (req, res) => {
   res.render('reports', { cssFile: ['reports'], jsFile: ['reports'] });
@@ -18,19 +21,40 @@ exports.get = (req, res) => {
 
 exports.post = (req, res) => {
   fixedVarible.findAll().then((fixedVaribleResult) => {
-    const variables = fixedVaribleResult[0].dataValues;
-    const variablesToCamel = convertToCamelCase(variables);
+    let variables = fixedVaribleResult[0].dataValues;
+    variables = convertToCamelCase(variables);
     bonusQuery().then((employeesAfterBonus) => {
       employeesAfterBonus.forEach((afterBonus) => {
-        deductionsQuery(convertToCamelCase(afterBonus), variablesToCamel).then(
+        afterBonus = convertToCamelCase(afterBonus);
+        deductionsQuery(afterBonus, variables).then(
           (afterDeductions) => {
-            let finalEmployee = Object.assign(afterBonus, afterDeductions);
-            finalEmployee = convertToCamelCase(finalEmployee);
-            const { totalDeductions, salary, totalAllownace } = finalEmployee;
-            const finalSalary = (salary + totalAllownace) - totalDeductions;
-            finalEmployee.finalSalary = finalSalary;
-            finalEmployee = convertToSnakeCase(finalEmployee);
-            console.log(finalEmployee);
+            exemptions(afterBonus, variables).then((employeeExemptions) => {
+              exemptionsOperations(
+                employeeExemptions,
+                afterDeductions,
+                variables,
+              ).then((finalExemptions) => {
+                deductionOperations(
+                  afterBonus,
+                  afterDeductions,
+                  finalExemptions,
+                  variables,
+                ).then((finalDeductions) => {
+                  let finalEmployee = Object.assign(
+                    afterBonus,
+                    finalExemptions,
+                    finalDeductions,
+                  );
+                  finalEmployee = convertToCamelCase(finalEmployee);
+                  console.log(finalEmployee);
+                });
+              });
+            });
+            // finalEmployee = convertToCamelCase(finalEmployee);
+            // const { totalDeductions, salary, totalAllownace } = finalEmployee;
+            // const finalSalary = (salary + totalAllownace) - totalDeductions;
+            // finalEmployee.finalSalary = finalSalary;
+            // finalEmployee = convertToSnakeCase(finalEmployee);
             // report.create(finalEmployee).then(() => {
             //   console.log(55);
             // });
@@ -38,36 +62,5 @@ exports.post = (req, res) => {
         );
       });
     });
-    // console.log(x);
-
-    //     employee.q ().then((employeeResult) => {
-    // employee.
-    //     })
-    //   const employeeResultRows = [];
-    //   employeeResult.forEach((employeeElement) => {
-
-    //     const convertedEmployeeElement = convertToCamelCase(
-    //       employeeElement.dataValues,
-    //     );
-    //     employeeResultRows.push(convertedEmployeeElement);
-    //   });
-    //   const finalReports = [];
-    //   employeeResultRows.forEach((employeeElement) => {
-    //   employeeElement.
-    // bonus(employeeElement).then((employeeAfterBonus) => {
-    //   deductions(employeeElement, employeeAfterBonus, variablesToCamel)
-    //     .then((employeeAfterDeductions) => {
-    //       finalReports.push(Object.assign(employeeAfterBonus, employeeAfterDeductions));
-    //     }).then(() => {
-    //       finalReports.forEach((element) => {
-    //         console.log(element);
-
-    //         const reportInSnakeCase = convertToSnakeCasesnake(element);
-    //         console.log(reportInSnakeCase);
-    //       });
-    //     });
-    // });
-    //   });
-    // });
   });
 };
